Move qheap to ext folder
