+++ packages/client/src/sync/fetcher/fetcher.ts
@@ -2,12 +2,12 @@ import debugDefault from 'debug'
 import { Readable, Writable } from 'stream'
 
 import { Event } from '../../types.js'
-import { Heap } from '../../util/qheap.js'
+import { Heap } from '../../ext/qheap.js'
 
 import type { Config } from '../../config.js'
 import type { Peer } from '../../net/peer/index.js'
 import type { PeerPool } from '../../net/peerpool.js'
-import type { QHeap } from '../../util/qheap.js'
+import type { QHeap } from '../../ext/qheap.js'
 import type { JobTask as BlockFetcherJobTask } from './blockfetcherbase.js'
 import type { Job } from './types.js'
 import type { Debugger } from 'debug'
